server/event-publisher/requirements.txt
pika==1.3.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.1